backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.28.0
diffusers>=0.29.0
transformers>=4.41.0
accelerate>=0.30.0
safetensors>=0.4.3
torch>=2.2.0
Pillow>=10.3.0
python-multipart>=0.0.9
